moved UriPatten class to 'core' package
