+++ components/uuf-core/src/test/java/org/wso2/carbon/uuf/ComponentTest.java
@@ -24,7 +24,7 @@ import org.wso2.carbon.uuf.core.Component;
 import org.wso2.carbon.uuf.core.Lookup;
 import org.wso2.carbon.uuf.core.Page;
 import org.wso2.carbon.uuf.core.RequestLookup;
-import org.wso2.carbon.uuf.internal.core.UriPatten;
+import org.wso2.carbon.uuf.core.UriPatten;
 import org.wso2.carbon.uuf.spi.model.Model;
 
 import java.util.Optional;
