+++ components/uuf-core/src/test/java/org/wso2/carbon/uuf/UriPattenTest.java
@@ -20,7 +20,7 @@ import com.google.common.collect.ImmutableMap;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-import org.wso2.carbon.uuf.internal.core.UriPatten;
+import org.wso2.carbon.uuf.core.UriPatten;
 
 import java.util.Map;
